quarter-2-openai-agents-sdk/class-07/tools/pyproject.toml
[project]
name = "tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.3.3",
    "python-dotenv>=1.1.1",
]
